+++ src/main/java/calculator/service/ServiceV1.java
@@ -21,7 +21,7 @@ public class ServiceV1 implements Service {
     }
 
 
-    public ArrayList<Integer> extractNumbers(String delimiter, NumberExtractor numberExtractor, String input) {
+    public ArrayList<Long> extractNumbers(String delimiter, NumberExtractor numberExtractor, String input) {
         NumberDto numberDTO = numberExtractor.extractNumbers(input, delimiter);
         return numberDTO.getNumberRepository();
 
@@ -43,7 +43,7 @@ public class ServiceV1 implements Service {
     }
 
 
-    public int addAll(ArrayList<Integer> extractNumbers) {
+    public long addAll(ArrayList<Long> extractNumbers) {
         return Calculator.allSum(extractNumbers);
     }
